client,app: Transaction History UI (#2666)
* client,app: Transaction History UI

This diff adds a UI to the wallets page displaying the transaction
history of a wallet.

Modifications are also made to the WalletTransaction type. The ID is now
a string rather than a []byte. `BalanceDelta int64` is now `Amount uint64`
as the sign is obvious from the type of the transaction and negatives are
a hassle to deal with. A `Recipient` field for `Send` transactions and a
`BondInfo` field for bond related transaction are also added.