enable tor
This is a proof-of-concept for enabling wallet access as a tor
hidden service. To enable the hidden service, run the build script,
client/tor/build_linux.sh, and rebuild dexc. The .onion address will
print in the logs during startup.

dexc must be run with the --tor flag.

This works by embedding the tor executable itself in the wallet binary,
and then writing it to file and running it when needed. This scheme is
intended to facilitate distribution, but other schemes are possible.

Websockets are not working and I don't know why.