client/asset/{dcr,btc}: RPC wallets TX History
This implements the transaction history functionality for DCR and BTC RPC
wallets. A separate DB is created for each wallet. The DBs are identified
by one of the addresses that the wallet controls. On connect, the wallet
will loop through each of the available DBs and see if the wallet controls
the address used to identify DB.