client/multi: Refactor MM Balance Handling (#2641)
* client/{asset,core,rpcserver}: Add WalletTransaction

- This adds a WalletTransaction function to the WalletHistorian interface.
It returns information about a transaction which the wallet made, or one
in which the wallet received funds. The TransactionConfirmations function
is removed, as it will no longer be needed once all wallets support
WalletTransaction.

- A `PartOfActiveBalance` field is added to the `WalletTransaction` type.
This is required for market making to know when it is safe to consider the
funds from an incoming transaction available for use.

- A `ParentAssetLockedAmt` field is added to the `core.Order` type. This is
needed in order to know the amount of token's fee asset is locked for an
order.

* client/mm: Refactor bot balance tracking

- This diff updates the balance tracking for market making bots. Previously,
there was a fragile diff based accounting technique. Now, on each bot
action, including trades, withdrawals, and deposits, a pending action is
stored. This pending action is updated until it is complete, at which time
it is removed, and the "base" balance of the bot is updated. The bot
balances are calculated by adding the base balances of the bot to the
effects of all pending actions.

- The `wrappedCore` and `wrappedCEX` types are removed, and their functionalities
are combined into a `unifiedExchangeAdaptor`. This is useful because
pending deposits and pending withdrawals have balance effects on both
CEX and DEX balances. Each `unifiedExchangeAdaptor` listens to a core
notification stream, and the `MarketMaker` type no longer needs to. In
a future refactor, the bots themselves will no longer listen to the
core notification functionality, and all common functionalities can be
moved to the `unifiedExchangeAdaptor`.

- Previously, the fee assets of tokens were not taken into account, but
now they are.

- The Binance library now uses a different endpoint, `/api/v3/balances`,
to get balances. This is supported by the Binance testnet api, so
testbinance is no longer required for balances.