mm: Live stats and archived event logs (#2711)
* mm: Live stats and archived event logs

This diff implements the ability to view the actions taken by market
maker bots and the performance of bots as they are running, and also
for already completed runs.

An event log db is created that is updated each time a bot initiates
a dex order, cex order, deposit, or withdrawal. It is also updated
each time one of these events has an update. The fiat rates of each
asset relevant to the market making bot is also tracked over the course
of the run to support a P/L calculation.

The unifiedExchangeAdaptor is updated to keep track of the effects of
a market making run, including the base/quote balance delta, and
base/quote fees. These values are updated whenever an event completes,
similar to how the balance tracking works. The P/L calculation for
running bots includes not only completed transactions, but also pending
transactions. If a swap is initiated, the expected redemption is also
included. Otherwise, while waiting for the redemtpion, it would appear
as if a loss had been incurred.

The UI is updated with an /mmarchives page which lists all of the archived
runs. From here, and from the main /mm page, a new /mmlogs page can be
reached, which displays each of the events that took place over the
course of the run.

* Buck review updates

* Store latest balance state instead of deltas