client/asset/{dcr,btc}: RPC wallets TX History (#2693)
* client/asset/{dcr,btc}: RPC wallets TX History

This implements the transaction history functionality for DCR and BTC RPC
wallets. A separate DB is created for each wallet. The DBs are identified
by one of the addresses that the wallet controls. On connect, the wallet
will loop through each of the available DBs and see if the wallet controls
the address used to identify DB.

- BadgerDB returns an ErrConflict when an update transaction encounters a
  stale read. These are now handled by retrying with an exponential backoff.
- The transaction databases now contain a WaitGroup that keeps track of
  the queries and garbage collection jobs. When `Close` is called, it
  waits for all jobs to finish before closing.