mm: Event log improvements (#2858)
* mm: Update pending event logs after run ends

When a market making run is ended, all orders and cancelled, but some of
these orders, deposits, and withdrawals may still be processing. This diff
updates  `MarketMaker.RunLogs` to check whether or not there are pending
events in the logs, and if so, update them with the latest state. The final
balance of the bot must also be updated to get correct P/L, so some
refactoring is done to simplify this. The `MarketMakingOverview` and
`MarketMakingEvents` are all updated to contain a `BalanceEffects` field.
When a pending event is updated, the final balances are updated by
the differences in the `BalanceEffects` of the old and new events. The
exchange adaptor is also updated to use `BalanceEffects` for each type of
event, making the code much simpler.

* app: Improve market making logs UI

Makes the UI look more similar to other pages, and also adds filtering and
pagination.

* CEX Withdrawal only pending if TxID is not available

* Fix balance tracking problems

* Fix mmlogs page back button

* Clear event log db if previous version