refactor tx sending
Refactor sending evm txs so that we store the txs before sending
and are accepting of send errors. This avoids a scenario where
we send the tx, but still receive an error, perhaps because of
a bad connection or other problem with the rpc provider.

Implement a send queue so that the caller doesn't have to wait
for the send. This should speed up ticks in core if the rpc
provider is functional but slow to respond.