client/comms,binance: Update ws connection URL
This updates the ws connection interface to allow for updating of the URL.
This is useful for the connection to Binance. If additional markets are
subscribed after the initial connection to the market streams API,
in the event of an error, the websocket connection will reconnect using
the initial URL, and the additional markets will only be resubscribed after
a list subscriptions check is done. With this change, all the markets will
be subscribed during a reconnection.