mm: Epoch reporting
This diff updates the market makers to generate a report of their
activities during each epoch. If the rates of each placement is unable
to be determined, this is reported as a `PreOrderProblem`, otherwise
an `OrderReport` is generated for the orders placed on each side of the
market containing information such as lots the bot is required to place,
the current number of lots standing, the number of lots booked, the
balances required and used for each placement, etc. This information is
all displayed on the UI.

A bug is also fixed in this diff. Previously, on a call to
`core.MultiTrade`, it was possible for some of the orders to be placed,
and some of them to result in an error. In this case, an error was returned
from `core.MultiTrade`, and the market makers would assume that none
of the trades were placed. Now a `core.MultiTradeResult` is returned for
each of the requested orders in the `MultiTrade` call containing either
the order or an error.