bolt: Improve delete inactive orders/matches perf.
Running these functions on a resource contrained machine and a large
database was incredibly slow due to overuse of the newestBuckets func.
Every key was accessed and sorted every time a batch of orders/matches
was processed.

Rather than repeatedly performing this work, the full set of keys is now
accessed only once. The resulting code runs ~200x faster.

There is a slight change in behaviour because database transactions are
now committed with a *maximum* of 1000 deletions rather than *exactly*
1000. This is not an issue because the exact size of each transaction
doesn't matter, the important point is to limit the maximum size of each
transaction to prevent excess memory consumption.

This performance gain removes the need to log the progress of each
individual batch, so now only a summary is logged after the entire
deletion process is completed.