mm: Handle market params update (#2846)
* mm: Handle market params update

This diff updates the market maker to listen for updates to the server
config. If the lot size of a market that a bot is currently trading on
is updated, the number of placements that the bot makes is updated so
that the bot places as many lots as possible without exceeding the amount
that was placed before the lot size update.

* Review updates

* Reivew updates