tatanka: Switch to AES for E2EE (#3204)
* tatanka: Switch to AES for E2EE

Replace RSA with AES for end-to-end encryption between mesh clients.
A permanent shared key is derived by hashing the result of an Elliptic
Curve Diffie-Hellman exchange using the permanent peer IDs of two peers.
This key secures the handshake, during which an ephemeral shared key is
negotiated. The ephemeral key encrypts all subsequent peer-to-peer
communication.

With all Tankagram communication now encrypted (using either the permanent
or ephemeral key), the unencrypted `Message` field is removed from
`Tankagram`.

Introduce `TankagramResult`, containing an error and payload. The payload
is encrypted and stored in the `EncryptedPayload` field of
`TankagramResult`. This provides a standardized way of communicating errors
in Tankagram communication.