firo: Electrum wallet get block height fix
The firo electrum wallet was unable to get the block height using the
hash because the code assumed that the block hash was calculated using
a double sha256 hash like Bitcoin, when in fact Firo has its own hash
algorithm. There is no golang library for FiroPOW, so instead of fetching
the block header that we are looking for and hashing it, we fetch the
subsequent block header and check the previous block hash field.

There were two almost identical syncTxHistory functions in the btc baseWallet
and the ExchangeWalletElectrum. The only difference was the block height
check. This logic is now passed in as a parameter, and the syncTxHistory
function in ExchangeWalletElectrum is removed.