mm: Add multi-hop strategy to arb market maker
Adds the ability to perform multi-hop arbitrage trades using the
arb market maker bot. If the exact market that exist on the DEX
does not exist on the CEX, but there exist two markets each
involving the base and quote assets of the DEX market and sharing
the same third asset, then the bot will perform arbitrage trades
by using both of these markets.

For example, if there is a DCR-BTC market on the DEX, the bot can
be configured to do multi-hop arbs on DCR-USDT and a BTC-USDT markets.
Then if a sell order is matched on a DEX, the bot will sell BTC on the
BTC-USDT market and buy DCR on the DCR-USDT market, and vice-versa for
buys.

Since two arb trades are being done, the chance of unfilled arb trades
becomes higher. In order to avoid the possibility of funds being stuck
in the intermediate asset, market trades are used rather than limit
trades.