lexi: Upgrade, add, and delete indexes
Adds the ability to upgrade, add, and delete indexes in the lexi db.
Instead of using the table.AddIndex function to create indexes, the
list of indexes are configured when the table is created. Since we
don't know the origial type of the data when it was put into the db,
when indexes are upgraded or new indexes are added, an indexing function
that takes bytes rather than lexi.KV must be used. The current indexing
function that takes lexi.KV is still allowed for indexes that are created
when the table is initially created for simplicity and in order to not
have to store all the data that is used in the index.