client/asset/eth: Bridge POL from POS to Ethereum (#3333)
* client/asset/eth: Bridge POL from POS to Ethereum

Implements bridging POL from Polygon POS to Ethereum. This uses an old
technology called plasma bridge which requires the user to submit two
completion transactions. The first one must be called by the user and
submits the transaction into an exit queue, and the second one may be
called by anyone, and clears the exit queue one by one. If the exit queue
is very long, the withdrawal we are interested in may not be processed. We
handle this by making one call to clear the exit queue, and if the exit
is not processed, we just wait until our exit is cleared.

Support for the POL erc20 token is also added. Previously we only supported
the old MATIC token.

* Joe review fixes